Add auto_composer command.
When auto_composer is used development mode is set based on the
environment Variable COMPOSER_DEV_MODE. If it's undefined the mode
isn't touched. If its anything else then 0 developer mode is enabled.

This can be used as post-[install|update]-cmd for composer to enable
development mode based on the mode composer was called with (--dev vs. --no-dev):
...
"scripts": {
    "development-auto": "zf-development-mode auto_composer",
    "post-install-cmd": ["@development-auto" ],
    "post-update-cmd": [ "@development-auto" ],
}
...