qa: revert formatting and schema location changes from phpunit.xml.dist
When you migrate the dist file, it sets the schema location to the canonical URL for it; this then adds latency when running PHPUnit, as it needs to fetch it before it can validate the configuration.
We can use the locally installed version still; we just need to update the structure to follow the schema.

Additionally, we tend to favor 4-space indentiation everywhere, so I reverted indentation changes.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>